+++ packages/data-table-server/src/dataTableService/index.ts
@@ -5,4 +5,5 @@
 
 export { DataTableServiceBuilder, getDataTableServiceType } from './DataTableServiceBuilder';
 export { DataTableService } from './DataTableService';
+export { getDataTableService } from './getDataTableService';
 export { DataTableParameter } from './types';
