+++ packages/data-table-server/src/middleware/attachDataTableToContext.ts
@@ -4,7 +4,8 @@
  */
 
 import { NextFunction, Request, Response } from 'express';
-import { DataTableServiceBuilder, getDataTableServiceType } from '../dataTableService';
+import { getDataTableService } from '../dataTableService';
+import { validatePermissions } from './helpers';
 
 /**
  * Finds the requested dataTable and attaches it to the context
@@ -16,30 +17,16 @@ export const attachDataTableToContext = async (
   next: NextFunction,
 ) => {
   try {
-    const { accessPolicy, models, params, ctx } = req;
+    const { models, params } = req;
     const { dataTableCode } = params;
     const dataTable = await models.dataTable.findOne({ code: dataTableCode });
     if (!dataTable) {
       throw new Error(`No data-table found with code ${dataTableCode}`);
     }
 
-    const permissionGroups = dataTable.permission_groups;
+    validatePermissions(dataTable, req);
 
-    if (
-      !(
-        permissionGroups.includes('*') ||
-        permissionGroups.some(permissionGroup => accessPolicy.allowsAnywhere(permissionGroup))
-      )
-    ) {
-      throw new Error(`User does not have permission to access data table ${dataTable.code}`);
-    }
-
-    const serviceType = getDataTableServiceType(dataTable);
-    const dataTableService = new DataTableServiceBuilder()
-      .setServiceType(serviceType)
-      .setContext({ apiClient: ctx.services, accessPolicy, models })
-      .setConfig(dataTable.config)
-      .build();
+    const dataTableService = getDataTableService(dataTable, req);
 
     req.ctx.dataTable = dataTable;
     req.ctx.dataTableService = dataTableService;
